requirements.txt
Django==1.9.8
pillow
